Feature

+++ app/src/main/java/com/music/musicStreamer/api/v1/client/PlaylistClient.java
@@ -55,9 +55,9 @@ public class PlaylistClient implements PlaylistGateway {
     }
 
     @Override
-    public PlaylistWithMusicEntity getById(final int id) {
+    public PlaylistWithMusicEntity getById(final Long id) {
         info(this.getClass(), "Get playlist by id");
-        final var playlistModel = playlistRepository.findById(id).orElseThrow(() -> new PlaylistException(PlaylistMessages.NOT_FOUND));
+        final var playlistModel = playlistRepository.findById(id.intValue()).orElseThrow(() -> new PlaylistException(PlaylistMessages.NOT_FOUND));
 
         info(this.getClass(), "Playlist found");
         info(this.getClass(), "Playlist id: " + playlistModel.getId());
@@ -68,11 +68,11 @@ public class PlaylistClient implements PlaylistGateway {
     }
 
     @Override
-    public List<PlaylistEntity> getByUserId(final int userId) {
+    public List<PlaylistEntity> getByUserId(final Long userId) {
         info(this.getClass(), "Get playlist by user id");
         info(this.getClass(), "User id: " + userId);
 
-        return playlistRepository.findAllByUserId(userId).stream().map(playlist -> {
+        return playlistRepository.findAllByUserId(userId.intValue()).stream().map(playlist -> {
             final var musicIds = playlistMusicGateway.getMusicIdByPlaylistId((long) playlist.getId());
             return playlistFactory.toEntity(playlist, musicIds);
         }).toList();
